+++ index.js
@@ -1,6 +1,7 @@
 const fs = require('fs');
 const { Command } = require('commander');
 const Transformer = require('./modules/transformer.js');
+const validation = require('./modules/validation.js');
 const { pipeline } = require('stream');
 
 const program = new Command();
@@ -30,19 +31,3 @@ pipeline(
     }
 );
 
-function validation(action, shift) {
-    if (action === undefined) {
-        console.error('action is required');
-        process.exit(1);
-    }
-
-    if (action !== 'encode' && action !== 'decode') {
-        console.error('Action (encode/decode) are required');
-        process.exit(1);
-    }
-
-    if (shift === undefined) {
-        console.error('Please, indicate shift');
-        process.exit(1);
-    }
-}
